Content-Length was not set to https MITM, fixes #9
When returning a MITM HTTP response, we cannot know for sure
how long is the response body (user might replaced the original
body) as a first aid, we'll use chunked encoding for all
responses. Further optimizations might include reading a small chunk
of the response, and if it ends soon - use Content-Length, or, let
the user set Content-Length himself. I still need to find a way
of finding out wether or not the response was changed, and in case
it wasn't, return the response verbatime.